change current- and potential-advisors to local variables
It's better to avoid creating turtle variable for things that
are only used as temporary variables inside local computations.

In this case, especially, the values of the turtle variables
were immediately getting out of sync with "reality", e.g.:
if a current advisor is removed, the link is severed but
the value of the current-advisors turtle variable is not
modified, thus being "out of date" until that turtle gets
selected again... and then the same thing happens.

In this case, I just moved code that used these variable inside the
evaluate-scenarios-and-choose procedure, but the variables could also
be passed as arguments to sub-procedures if preferred.